Linux: merged compiler warning suppression as well as duplicated symbol bug on linux
git-svn-id: file:///dev/shm/opensim/Branches/OpenSim30@7734 fc29be9c-2cb0-420d-b520-73dbd18add6e

+++ OpenSim/Actuators/Thelen2003Muscle.cpp
@@ -436,7 +436,7 @@ void Thelen2003Muscle::computeInitialFiberEquilibrium(SimTK::State& s) const
     
         int flag_status    = (int)soln[0];
         double solnErr        = soln[1];
-        double iterations     = (int)soln[2];
+        int iterations     = (int)soln[2];
         double fiberLength    = soln[3];
         double passiveForce   = soln[4];
         double tendonForce    = soln[5];
